feat(customElement): Add bindable property change event handlers
Use the `propertyChanged` method called by Aurelia for every prop but
the icon. There can be some optimizations added later, such as adding a
`maskChanged` or `transformChanged`, since these need to run additional logic,
but it seems minimal right now.

When the icon changes it needs to be normalized before it is rendered,
therefore it has to call the `attached` method first in the
`iconChanged` method.

closes #5