refactor(customElement): Bind icon html to innerhtml
Use `innerhtml.bind` to change the icon when a property changes. The
slot and view do not seem to be needed since font awesome and the
private convert method do all the work of creating the SVGElement from
the view model's bound properties. The custom element only needs the new
html