feat(plugin): Add visitor plugin configuration
PR #21: Use visitor plugin to add default bindable values that all icons are
initiated with, which are passed in when aurelia starts. Hold reference to this
object so future updates are used during runtime. Update devapp to look
modern. Visitor must be used in ctor or else it could get in the way of the lifecycle.

closes #11 
related #18